bot-requirements.txt
numpy==1.20.1
Cython==0.29.21
pandas==1.2.3
requests==2.22.0
scipy==1.6.0
future==0.18.2
google-cloud-storage==1.36.0
google-cloud-bigquery==2.9.0
google-cloud-bigquery-storage==2.3.0
pyarrow==3.0.0
tzdata==2021.1
six==1.15.0
backports.ssl_match_hostname==3.7.0.1
websocket-client==0.57.0
PyPortfolioOpt==1.4.1
alpaca-trade-api>=0.25
discord.py==1.6.0
ta